discovery/skillsmp_scraper.py
"""
SkillsMP scraper - discovers top skills from skillsmp.com
"""

import argparse
import json
import time
from pathlib import Path
from typing import List, Dict
import requests
from bs4 import BeautifulSoup


class SkillsMPScraper:
    """Scrapes skill information from skillsmp.com"""

    def __init__(self, base_url: str = "https://skillsmp.com"):
        """
        Initialize scraper.

        Args:
            base_url: Base URL for SkillsMP (default: https://skillsmp.com)
        """
        self.base_url = base_url
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36'
        })

    def scrape_skills_page(self, page: int = 1) -> List[Dict]:
        """
        Scrape skills from a single page.

        Args:
            page: Page number to scrape

        Returns:
            List of skill dictionaries with name, url, description
        """
        # Note: This is a placeholder implementation
        # The actual URL structure and HTML selectors would need to be
        # determined by inspecting skillsmp.com

        print(f"Scraping page {page}...")

        try:
            url = f"{self.base_url}/skills?page={page}"
            response = self.session.get(url, timeout=10)
            response.raise_for_status()

            soup = BeautifulSoup(response.text, 'html.parser')

            skills = []

            # TODO: Update selectors based on actual site structure
            # This is a generic implementation
            skill_elements = soup.find_all('div', class_='skill-card')

            for element in skill_elements:
                try:
                    name = element.find('h3').text.strip()
                    description = element.find('p', class_='description').text.strip()
                    link = element.find('a', href=True)
                    github_url = link['href'] if link else None

                    skills.append({
                        'name': name,
                        'description': description,
                        'github_url': github_url,
                        'source': 'skillsmp.com',
                        'page': page
                    })
                except Exception as e:
                    print(f"  Error parsing skill element: {e}")
                    continue

            print(f"  Found {len(skills)} skills")
            return skills

        except requests.RequestException as e:
            print(f"  Error fetching page {page}: {e}")
            return []

    def scrape_multiple_pages(self, num_pages: int = 1) -> List[Dict]:
        """
        Scrape multiple pages of skills.

        Args:
            num_pages: Number of pages to scrape

        Returns:
            Combined list of all skills
        """
        all_skills = []

        for page in range(1, num_pages + 1):
            skills = self.scrape_skills_page(page)
            all_skills.extend(skills)

            # Be respectful with rate limiting
            if page < num_pages:
                time.sleep(1)

        return all_skills

    def save_skills(self, skills: List[Dict], output_path: str = "data/discovered/skills.json"):
        """
        Save discovered skills to JSON file.

        Args:
            skills: List of skill dictionaries
            output_path: Path to save JSON file
        """
        output_file = Path(output_path)
        output_file.parent.mkdir(parents=True, exist_ok=True)

        with open(output_file, 'w') as f:
            json.dump(skills, f, indent=2)

        print(f"\nSaved {len(skills)} skills to: {output_path}")


def create_mock_skills() -> List[Dict]:
    """
    Create mock skill data for testing when scraping isn't possible.

    Returns:
        List of mock skill dictionaries
    """
    return [
        {
            'name': 'ms-office-suite',
            'description': 'Create and manipulate Microsoft Office files (Excel, Word, PowerPoint)',
            'github_url': 'https://github.com/example/ms-office-suite-skill',
            'source': 'mock',
            'page': 1
        },
        {
            'name': 'pdf',
            'description': 'Create, read, and manipulate PDF documents',
            'github_url': 'https://github.com/example/pdf-skill',
            'source': 'mock',
            'page': 1
        },
        {
            'name': 'web-scraper',
            'description': 'Extract data from websites',
            'github_url': 'https://github.com/example/web-scraper-skill',
            'source': 'mock',
            'page': 1
        },
        {
            'name': 'data-viz',
            'description': 'Create data visualizations and charts',
            'github_url': 'https://github.com/example/data-viz-skill',
            'source': 'mock',
            'page': 1
        },
        {
            'name': 'email-sender',
            'description': 'Send emails via SMTP',
            'github_url': 'https://github.com/example/email-sender-skill',
            'source': 'mock',
            'page': 1
        }
    ]


def main():
    """CLI entry point"""
    parser = argparse.ArgumentParser(
        description="Scrape skills from SkillsMP"
    )

    parser.add_argument(
        '--pages',
        type=int,
        default=1,
        help="Number of pages to scrape (default: 1)"
    )

    parser.add_argument(
        '--output',
        type=str,
        default="data/discovered/skills.json",
        help="Output file path (default: data/discovered/skills.json)"
    )

    parser.add_argument(
        '--mock',
        action='store_true',
        help="Use mock data instead of scraping (for testing)"
    )

    args = parser.parse_args()

    if args.mock:
        print("Using mock skill data...")
        skills = create_mock_skills()
    else:
        scraper = SkillsMPScraper()
        skills = scraper.scrape_multiple_pages(args.pages)

        # If scraping failed, fall back to mock data
        if not skills:
            print("\nScraping failed. Using mock data as fallback...")
            skills = create_mock_skills()

    # Save results
    scraper = SkillsMPScraper()
    scraper.save_skills(skills, args.output)

    print(f"\nDiscovered {len(skills)} skills:")
    for skill in skills:
        print(f"  - {skill['name']}: {skill['description'][:60]}...")


if __name__ == "__main__":
    main()
