requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
python-pptx>=0.6.0
tiktoken>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
